Comments added + other bug fixes

+++ loadpath3D.m
@@ -1,15 +1,17 @@
 function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
                     overlay, parallel, newPDF, recompute, step_size, path_length)    
-%% ********************  Seed Coordinates  ******************************
-    F = findall(0,'type','figure','tag','TMWWaitbar'); %%%Unsupress this
-%     if wait bar wont close
+%% ********************  House Keeping   ******************************
+% Closes previously opened waitbars
+    F = findall(0,'type','figure','tag','TMWWaitbar'); 
     delete(F);
+
+    % Read's seed data in
     Seed = importdata(seed_dir, ',');
     xseed = Seed(:,1);
     yseed = Seed(:,2);
     zseed = Seed(:,3);
     numSeeds = length(xseed);
-%% HouseKeeping
+%% HouseKeeping - waitbar setup
     
     wb = waitbar(0,'1','Name','Computing Load Paths',...
             'CreateCancelBtn',...
@@ -24,52 +26,71 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
     warning('off','MATLAB:scatteredInterpolant:DupPtsAvValuesWarnId');
     warning('off','MATLAB:MKDIR:DirectoryExists')
 
-    %% Scan Data
+    %% Naming output files and killing interfering processes
     
     model_name = [model_name, ' - ',upper(path_dir), ' Path'];
     
     model_data_name = regexprep(model_name, ' ', '_');
+    %Want to make this platform independent. This line is only supported in
+    %windows distributions. Working on generalising, as yest dont have
+    %corresponding mac command.
+%     system(['taskkill /fi "WINDOWTITLE eq ', model_name,'.pdf"']);
+    if ismac
+        slash = '/';
+       
+    elseif ispc
+        slash = '\';
+        system(['taskkill /fi "WINDOWTITLE eq ', model_name,'.pdf"']);
+       
+    end
     
-    system(['taskkill /fi "WINDOWTITLE eq ', model_name,'.pdf"']);
-
-    nodei = [sim_dir '\' 'nodeInfo.txt'];
-
+    nodei = [sim_dir slash 'nodeInfo.txt'];
     numNodes = importdata(nodei);
     numNodes = numNodes(2);
 
     
     %% ******************  Populate Nodes and Elements  *********************
     
-    if ~exist([save_dir, '\Path Data\data_',model_data_name,'.mat'], 'file') || recompute
+    % Detects whether previous data has been computed, if yes, skips
+    % recomputation unless forced by user in GUI
+
+    if ~exist([save_dir slash 'Path Data' slash 'data_' model_data_name,'.mat'], 'file') || recompute
+
         fprintf('New model or user nominated to recompute data. Starting now.\n')
         waitbar(current_time/total_time,wb,sprintf('Computing Initial Data'))
         
+        %Nodal Information module
         [StressData, numNodes] = nodeDat(sim_dir, numNodes);
         current_time = current_time + data_read_time/3;
         waitbar(current_time/total_time,wb,sprintf('Computing Initial Data'))
         fprintf('Nodal information complete. Starting stress population.\n')
         
+        %Node data module
         [nodes] = NodeDatRead(sim_dir, StressData, numNodes);
         current_time = current_time + data_read_time/3;
         waitbar(current_time/total_time,wb,sprintf('Computing Initial Data'))
         fprintf('Nodal stresses populated. Element generation beginning.\n')
         
-        [MeshNode, nodePerEl, PartArr] = datread(sim_dir, nodes); 
+        %Element data and main data structure generation
+        [nodePerEl, PartArr] = datread(sim_dir, nodes); 
         current_time = current_time + data_read_time/3;
         
         fprintf('Elements constructed, directories being created and data being saved.\n')
-        mkdir(save_dir, '\Path Data')
-        save([save_dir,'\Path Data\data_',model_data_name,'.mat'],'PartArr','nodes', 'nodePerEl', 'MeshNode');
+        mkdir([save_dir, slash 'Path Data'])
+        save([save_dir,slash 'Path Data' slash 'data_',model_data_name,'.mat'],'PartArr','nodes', 'nodePerEl');
         
     else
+        %This loads data if the preprocessign has already been done.
         fprintf('Previous model detected, loading data.\n')
         waitbar(current_time/total_time,wb,sprintf('Loading Data'))
-        load([save_dir, '\Path Data\data_',model_data_name,'.mat']);
+        load([save_dir slash 'Path Data' slash 'data_' model_data_name,'.mat']);
         current_time = current_time + data_read_time;
         
         fprintf('Data loaded. Starting path computation.\n')
         
     end
+        %******************** Waitbar and Status Update ***************************
+
     if getappdata(wb,'canceling')
         delete(wb)
         return
@@ -78,7 +99,7 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
     waitbar(current_time/total_time,wb,sprintf('Starting Paths'))
     %% ****************  Load Path Generation  ******************************
     
-    
+    %Initialise data containers for load paths
     
     Paths(numSeeds).X.forward = [];
     Paths(numSeeds).Y.forward = [];
@@ -89,20 +110,27 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
     Paths(numSeeds).Z.total = [];
     Paths(numSeeds).I.total = [];
         
+    % Not happy with the repeated code below - needs to be cleaned.    
     switch parallel
+        %Parallel computation if load paths
         case 1
             workers = 4;
             switch dim
+                % Currently 2D and 3D are separate, very crude. Future
+                % update is to pass as vector and scall all functions
+                % according to the length of that vector.
                 case {'3D'}
                     parfor (i = 1:numSeeds, workers)
                         fprintf('Starting path %i\n',i)
                         warning('off','MATLAB:scatteredInterpolant:DupPtsAvValuesWarnId');
+                        %Main work horse module - Runge Kutta
                         [x, y, z, intense] = rungekuttaNatInter3D(xseed(i),...
                             yseed(i),zseed(i), PartArr, path_dir, nodePerEl,path_length,false,step_size,wb);
                         if isempty(x)
                             fprintf('Path %i unsuccessful\n',i)
                             continue
                         end
+
                         Paths(i).X.forward = x;
                         Paths(i).Y.forward = y;
                         Paths(i).Z.forward = z;
@@ -215,14 +243,28 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
             end
     end        
     fprintf('All seeds tested and appropriate paths computed. Saving path data.\n')
-    %% ****************  Plotting and Printing  ******************************
+
+    %******************** Waitbar and Status Update ***************************
+
+
     if getappdata(wb,'canceling')
         delete(wb)
         return
     end
     waitbar(current_time/total_time,wb,sprintf('Paths Finished Paths, Saving and Plotting now...\n'))
 
-    save([save_dir,'\Path Data\pathdata_',model_data_name,'.mat'], 'Paths');
+    %% ****************  Plotting and Printing  ******************************
+    %Data is output to .mat files so that in a future update the user can
+    %modulate the load path program. For example they could load previous
+    %paths and just run the plotting and printing section of the code. Or
+    %the user could just compute the paths and then send them to someone
+    %else to plot them on their machine or with their specific settings.
+
+    %Other considerations are for future transient analysis where multiple
+    %data sets may have to be condensed. And that its a good backup of the
+    %path calculation.
+
+    save([save_dir slash 'Path Data' slash 'pathdata_' model_data_name '.mat'], 'Paths');
     fig = figure;
     fprintf('Plotting Paths\n')
     
@@ -233,7 +275,7 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
             modelPlot2D([Paths(:).X],[Paths(:).Y],[Paths(:).I],PartArr)
     end
     
-    
+    %******************** Waitbar and Status Update ***************************
     if getappdata(wb,'canceling')
         delete(wb)
         return
@@ -242,27 +284,35 @@ function loadpath3D(dim, sim_dir, seed_dir, save_dir, model_name,path_dir,...
     current_time = current_time + plot_time;
     waitbar(current_time/total_time,wb,sprintf('Printing PDF'))
     
-    mkdir(save_dir,'\Path Plots')
+    % Create new directory to store the output plots
+    mkdir(save_dir,[slash 'Path Plots'])
     if newPDF
         dt = datestr(now,'HH.MM.SS_dd/mm/yy');
-        dateAppenedFN = [save_dir,'\Path Plots\',model_name,'_', dt, '.pdf'];
+        dateAppenedFN = [save_dir, slash 'Path Plots' slash ,model_name,'_', dt, '.pdf'];
     else
-        dateAppenedFN = [save_dir,'\Path Plots\',model_name, '.pdf'];
+        dateAppenedFN = [save_dir,slash 'Path Plots' slash ,model_name, '.pdf'];
     end
     
-
+    % Contrary to variable name and the description in the GUI, this was
+    % repurposed to let the user choose where to print the plot to a pdf or
+    % not. As the density is so high to see the paths properly it can be an
+    % expensive task to compute.
+    
+    %Future update will check whether an instance of the .pdf is already
+    %open and modify the name so that saving conflicts dont happen.
     switch overlay
         case 1
             gcf        
             print(fig,dateAppenedFN, '-dpdf','-r2000', '-fillpage');
             open(dateAppenedFN);
         case 0
-            hold off
+            
     end
     delete(wb)
     fprintf('Load paths complete.\n')
 end
 function [] = modelPlot3D(x_paths,y_paths,z_paths,Intensity,PartArr)
+    %Just some custom settings for plotting the paths
     Alpha = 0.1;
     Buffer = 0.35;
     wireFrame(PartArr,Alpha, Buffer);
@@ -288,6 +338,7 @@ function [] = modelPlot3D(x_paths,y_paths,z_paths,Intensity,PartArr)
     colorbar;
 end
 function [] = modelPlot2D(x_paths,y_paths,Intensity,PartArr)
+%Just some custom settings for plotting the paths
     Alpha = 0.1;
     Buffer = 0.35;
     wireFrame2D(PartArr,Alpha, Buffer);
